requirements.txt
ultralytics>=8.0.0
opencv-python
numpy
pandas
matplotlib
seaborn
streamlit
python-dotenv
